+++ Tools/OSD/Constants.cs
@@ -84,12 +84,12 @@ namespace OSD {
 		
 	// new!
 		public const int useExtVbattA			= 4;
-		public const int useExtVbattB			= 5;
+		public const int useExtVbattB 			= 5;
     	public const int useExtCurr				= 6;
 		public const int AutoScreenSwitch       = 7;
 		public const int results_on     		= 8;
 		public const int mode_auto              = 9;
-        public const int flgHUD                 = 10;
+        public const int battB_is_on            = 10;
         public const int flgTrack               = 11;